fix: remove ::numeric cast on confidence_score (column is text, values like "high")

+++ supabase/migrations/20260630000005_fix_pup_unique_key_composite.sql
@@ -109,7 +109,7 @@ BEGIN
     ),
     (attendance.value->>'join_time')::timestamptz,
     (attendance.value->>'leave_time')::timestamptz,
-    (attendance.value->>'confidence_score')::numeric
+    attendance.value->>'confidence_score'
   FROM jsonb_array_elements(new_attendance_data) AS attendance
   LEFT JOIN prickle_id_map
     ON prickle_id_map.client_prickle_id = attendance.value->>'client_prickle_id'
